pyproject.toml
[project]
name = "Flashcard-study-app"
version = "0.1.0"
description = ""
readme = ""
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "markdown>=3.10.2",
    "posit>=0.1.2",
    "posit-sdk>=0.16.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "rsconnect-python>=1.28.2",
]
